Develop

+++ app/src/main/java/com/jmm/portableairquality/Model/CcsMeasurement.java
@@ -3,8 +3,8 @@ package com.jmm.portableairquality.Model;
 import java.io.Serializable;
 
 public class CcsMeasurement implements Serializable {
-    public long co2;
-    public long voc;
+    public int co2;
+    public int voc;
 
     public CcsMeasurement(byte[] value) {
         co2 = (value[3] & 0xFF) << 8 | value[2] & 0xFF;
